ENH: split get_tasks and Task base class into two files, setup the object oriented calls that will be made for each task (as discussed on slack)
